requirements.txt
Flask==0.12.2
flask-marshmallow==0.8.0
Flask-SQLAlchemy==2.2
PyMySQL==0.7.11
SQLAlchemy==1.1.13
slackweb==1.0.5
marshmallow_enum=1.4.1